ANDROID-11907 Remove tests for PreviewPasswordVisibleIcon
